requirements.txt
slack-sdk
slack-bolt
# Flask
fastapi
uvicorn
python-dotenv
langchain
openai